feat: add `:SqlsExecuteQueryVertical`
It turns out I misunderstood the purpose that command in the original
`sqls.vim` plugin. I thought it opened a Vim window vertically and thus
opted to use modifiers like `:vertical` instead. The true purpose of
that command is to display the results of the query vertically and is
not related to the position of the preview window at all.